test(url-key): give url-key.mjs its own suite
ARCHITECTURE.md and CONTRIBUTING.md both specify tests/{module}.test.mjs for
the root script a suite covers. url-key.mjs had no suite at all: its unit cases
lived inside tests/merge-tracker-url-dedup.test.mjs, so the only way to run them
was to run a suite named for a different script, and anyone changing
normalizeUrl had no obvious file to reach for.

Pure move, no behaviour change. The file was already sectioned by a divider
comment, and the split follows it exactly: 6 normalizeUrl blocks out, 16
merge-tracker integration blocks stay. Nothing exercised both. The two headers
now point at each other, and the unused import is dropped from the file that no
longer needs it.

test-all.mjs discovers tests/**/*.test.mjs with no registration list, so the new
suite runs from the moment it exists.

Closes #4070
